+++ CHANGELOG.md
@@ -17,7 +17,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - Make model internal properties private when possible.
 - The package no longer depends on `infuse/libs`.
 - Renamed `Property` class to `Type`.
-- A new `Definition` class contains all the properties that belong to a method. Retrieving the model properties will return this object instead of an array.
+- A new `Definition` class contains all the properties that belong to a method. Retrieving the model properties with `Model::definition()` will return this object instead of an array.
 - A new `Property` class encapsulates the definition of a model property. The model property methods will return this object instead of an array.
 - Moved `Model::cast()` to `Type::cast()`
 - Marked classes as final when appropriate.
@@ -26,6 +26,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 - The unique constraint is now a validation rule specified as: `['validate' => ['unique', 'column' => 'email']]`.
 - Deprecated the `password_php` validation rule in favor of `password`.
 - Callable validation rules must now be specified as: `['validate' => ['callable', 'fn' => ...]]`.
+- Deprecated `Model::getProperties()`, `::hasProperty()`, and `::getProperty()`.
 
 ### Fixed
 - Use a strict equality check when casting an empty string to null on a nullable property. Previously this would check for a falsey value.